pyproject.toml
[project]
name = "device-mcp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastmcp>=3.0.2",
    "netmiko>=4.6.0",
    "pyyaml>=6.0.3",
    "anthropic>=0.40.0",
    "mcp>=1.0.0",
    "python-dotenv>=1.0.0",
]
